+++ Sources/SwiftUIKit/_Deprecated/ActionButton.swift
@@ -1,8 +1,0 @@
-#if os(iOS)
-@available(iOS 14, *)
-@available(*, deprecated, renamed: "FormActionButton")
-public typealias ActionButton = FormActionButton
-
-@available(*, deprecated, renamed: "FormActionButtonStyle")
-public typealias ActionButtonStyle = FormActionButtonStyle
-#endif