+++ Sources/SwiftUIKit/_Deprecated/MenuListItemIcon.swift
@@ -1,4 +1,0 @@
-import SwiftUI
-
-@available(*, deprecated, renamed: "MenuListIcon")
-typealias MenuListItemIcon = MenuListIcon