+++ Sources/SwiftUIKit/_Deprecated/UserDefaultsPersisted.swift
@@ -1,2 +1,0 @@
-@available(*, deprecated, renamed: "Persisted")
-public typealias UserDefaultsPersisted = Persisted