+++ Sources/SwiftUIKit/Colors/SystemColor.swift
@@ -23,7 +23,7 @@ import SwiftUI
  the current platform. Avoid this by filtering out supported
  values with the `isSupported` property.
  */
-public enum SystemColor: String, CaseIterable {
+public enum SystemColor: String, CaseIterable, Codable, Equatable {
     
     case primary
     case secondary