+++ Sources/SwiftUIKit/Views/MenuList/MenuListToggle.swift
@@ -1,5 +1,5 @@
 //
-//  MenuListItemStack.swift
+//  MenuListToggle.swift
 //  SwiftUIKit
 //
 //  Created by Daniel Saidi on 2021-02-11.