+++ Release Notes.md
@@ -8,7 +8,9 @@ This version requires Xcode 13 and later, since it refers to the latest api:s.
 ### ✨ New features
 
 * `Color+Codable` is a new extension that makes Color codable.
-* `EnvironmentValues+SafeAreaInsets` lets you read the safe are insets of the key window.  
+* `EnvironmentValues+SafeAreaInsets` lets you read the safe are insets of the key window.
+* `FormActionButton` can now be provided with a custom style.  
+* `FormActionButtonStyle` can now be configured with a custom style.  
 * `MenuListIcon` has a new emoji-based initializer.
 * `MenuListItem` now lets you use any custom content builder.
 * `MenuListSubtitle` is a new view for menu list subtitles.